uapi: convert fdset.h
