+++ include/protura/fs/fdset.h
@@ -8,42 +8,7 @@
 #ifndef INCLUDE_PROTURA_FS_FDSET_H
 #define INCLUDE_PROTURA_FS_FDSET_H
 
-#ifdef __KERNEL__
-# include <protura/types.h>
-# include <arch/bitops.h>
-#endif
-
-#define __kNOFILE 32
-#define __kFD_SETSIZE 32
-
-typedef unsigned int __kfd_set;
-typedef unsigned int __kfd_mask;
-
-#define __kNBBY 8 /* Number of bits in a byte */
-
-#define __kNFDBITS (sizeof(fd_mask) * NBBY)
-
-static inline void __kFD_CLR(int fd, __kfd_set *set)
-{
-    *set &= ~(1 << fd);
-}
-
-static inline void __kFD_SET(int fd, __kfd_set *set)
-{
-    *set |= (1 << fd);
-}
-
-static inline int __kFD_ISSET(int fd, __kfd_set *set)
-{
-    return !!(*set & (1 << fd));
-}
-
-static inline void __kFD_ZERO(__kfd_set *set)
-{
-    *set = 0;
-}
-
-#ifdef __KERNEL__
+#include <uapi/protura/fs/fdset.h>
 
 #define NOFILE __kNOFILE
 #define FD_SETSIZE __kFD_SETSIZE
@@ -58,7 +23,5 @@ typedef __kfd_mask fd_mask;
 #define FD_SET(fd, set) __kFD_SET(fd, set)
 #define FD_ISSET(fd, set) __kFD_ISSET(fd, set)
 #define FD_ZERO(set) __kFD_ZERO(set)
-
-#endif
 
 #endif
